requirements.txt
# Core
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0

# OCR & Image Processing
opencv-python>=4.8.0
Pillow>=10.1.0
easyocr>=1.7.0
numpy>=1.26.0
deepface>=0.0.89
tf-keras>=2.16.0

# PDF Support
pdf2image>=1.16.0
PyPDF2>=3.0.0

# Database
sqlalchemy>=2.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0

# Rate Limiting
slowapi>=0.1.9

# Utilities
aiofiles>=23.2.0
python-dotenv>=1.0.0
requests>=2.31.0
loguru>=0.7.0
tenacity>=8.2.0

# Scientific (pulled by easyocr, listed for clarity)
scipy>=1.11.0
scikit-image>=0.22.0
scikit-learn>=1.3.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
